refactor:도메인 계층 패키지 구조 변경
